Move helpers into 0x file
